+++ empire/server/v2/api/bypass/bypass_dto.py
@@ -3,12 +3,14 @@ from typing import List
 
 from pydantic import BaseModel
 
+from empire.server.v2.api.shared_dto import Author
 
-# TODO VR Authors
+
 def domain_to_dto_bypass(bypass):
     return Bypass(
         id=bypass.id,
         name=bypass.name,
+        authors=bypass.authors or [],
         code=bypass.code,
         created_at=bypass.created_at,
         updated_at=bypass.updated_at,
@@ -18,6 +20,7 @@ def domain_to_dto_bypass(bypass):
 class Bypass(BaseModel):
     id: int
     name: str
+    authors: List[Author]
     code: str
     created_at: datetime
     updated_at: datetime